requirements.txt
kafka-python==2.0.2
flask==2.2.2
jinja2==3.1.2
markupsafe==2.1.1
werkzeug==2.2.2
click==8.1.3
colorama==0.4.5
itsdangerous==2.1.2
greenlet==1.1.3
SQLAlchemy==1.4.41
Flask-SQLAlchemy==2.5.1
dominate==2.7.0
visitor==0.1.3
pandas==1.4.4
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
joblib==1.2.0
scikit-surprise==1.1.1
scipy==1.9.1
surprise==0.1
scikit-learn==1.1.2 
sklearn==0.0 
requests==2.28.1
threadpoolctl==3.1.0
contourpy==1.0.5 
cycler==0.11.0 
fonttools==4.37.2 
kiwisolver==1.4.4 
matplotlib==3.6.0 
packaging==21.3 
pillow==9.2.0 
pyparsing==3.0.9
attrs==22.1.0
exceptiongroup==1.0.0rc9
iniconfig==1.1.1
pluggy==1.0.0
pytest==7.2.0
tomli==2.0.1
pytest-mock==3.10.0
pytest-cov==4.0.0
certifi==2022.9.24
cloudpickle==2.2.0 
idna==3.4 
protobuf==3.20.3 
pytimeparse==1.1.8 
pyyaml==6.0 
requests==2.28.1 
urllib3==1.26.12 
verta==0.21.0
prometheus_client==0.15.0
dvc==2.35.2